+++ raiden-ts/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+## [0.14.0] - 2020-11-25
 ### Fixed
 - [#2360] Properly error & shutdown if database gets deleted at runtime
 
@@ -367,7 +369,8 @@
 - Add protocol message implementation.
 
 
-[Unreleased]: https://github.com/raiden-network/light-client/compare/v0.13.0...HEAD
+[Unreleased]: https://github.com/raiden-network/light-client/compare/v0.14.0...HEAD
+[0.14.0]: https://github.com/raiden-network/light-client/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/raiden-network/light-client/compare/v0.12.0...v0.13.0
 [0.12.0]: https://github.com/raiden-network/light-client/compare/v0.11.1...v0.12.0
 [0.11.1]: https://github.com/raiden-network/light-client/compare/v0.11.0...v0.11.1